+++ nd4j/nd4j-backends/nd4j-api-parent/nd4j-api/src/main/java/org/nd4j/autodiff/samediff/internal/InferenceSession.java
@@ -16,6 +16,9 @@
 
 package org.nd4j.autodiff.samediff.internal;
 
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
 import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
 import org.nd4j.autodiff.functions.DifferentialFunction;
@@ -26,7 +29,6 @@ import org.nd4j.autodiff.samediff.SameDiff;
 import org.nd4j.autodiff.samediff.VariableType;
 import org.nd4j.autodiff.samediff.internal.memory.SimpleSessionMemoryMgr;
 import org.nd4j.base.Preconditions;
-import org.nd4j.collections.WeakIdentityHashMap;
 import org.nd4j.linalg.api.buffer.DataType;
 import org.nd4j.linalg.api.memory.MemoryWorkspace;
 import org.nd4j.linalg.api.ndarray.INDArray;
@@ -45,6 +47,7 @@ import org.nd4j.linalg.indexing.INDArrayIndex;
 import org.nd4j.linalg.indexing.NDArrayIndex;
 import org.nd4j.linalg.util.ArrayUtil;
 
+import java.io.LineNumberReader;
 import java.util.*;
 
 /**
@@ -59,20 +62,13 @@ public class InferenceSession extends AbstractSession<INDArray,SameDiffOp> {
             "Alternatively, arrays defined in a workspace must be replaced after the workspace has been closed.";
 
     private SessionMemMrg mmgr;
+    private DependencyTracker<Array,Dep> arrayUseTracker = new DependencyTracker<>();       //What needs to happen before the array can be closed?
 
-    /**
-     * A weak identity hash set
-     */
-    private Set<INDArray> identitySetAllConstPhVar;
 
     public InferenceSession(@NonNull SameDiff sameDiff) {
         super(sameDiff);
 
         mmgr = new SimpleSessionMemoryMgr();
-
-
-
-        identitySetAllConstPhVar = Collections.newSetFromMap(new WeakIdentityHashMap<INDArray, Boolean>());
     }
 
     @Override
@@ -115,6 +111,7 @@ public class InferenceSession extends AbstractSession<INDArray,SameDiffOp> {
             }
             out.put(e.getKey(), arr);
         }
+
         return out;
     }
 
@@ -132,9 +129,6 @@ public class InferenceSession extends AbstractSession<INDArray,SameDiffOp> {
                 clearOpReferencesFor(s);
             }
         }
-
-        //Also clear identity set, in case user has changed any arrays since last run
-        identitySetAllConstPhVar.clear();
 
         return output;
     }
@@ -178,216 +172,323 @@ public class InferenceSession extends AbstractSession<INDArray,SameDiffOp> {
         }
 
         /*
-        Deallocate input arrays if possible
-        We can only do this in the following cases:
-        1. The input is an ARRAY type SDVariable (not placeholder, constant, variable)
-        2. The input is not one of the requested output arrays
-        3. The input has been "fully consumed" - i.e., isn't required to calculate any other required ops
+        Check if any arrays can be deallocated.
+        We use array dependency tracking for this
          */
-        List<String> inputs = op.getInputsToOp();
-        if(inputs != null && !inputs.isEmpty()){
-            for( int i=0; i<inputs.size(); i++ ) {
-                String inName = inputs.get(i);
 
-                SDVariable v = sameDiff.getVariable(inName);
-                if(v.getVariableType() != VariableType.ARRAY || allReqVariables.contains(inName)){
+        //Step 1: Update dependencies - add new dependencies
+        List<String> opOutputs = op.getOutputsOfOp();
+        recordArrayUses(opOutputs, outputFrameIter, opInputs, allIterInputs, constAndPhInputs);
+        /*
+        protected void recordArrayUses(List<String> forVariables, FrameIter outputFrameIter, Set<VarId> opInputs, Set<VarId> allIterInputs,
+                                   Set<String> constAndPhInputs){
+         */
+//        for( int i=0; i<opOutputs.size(); i++ ){
+//            //Add the dependencies: What has to be executed, before we can close this just calculated array?
+//            String outVarName = opOutputs.get(i);
+//            List<String> thisOutputUsedForOps = sameDiff.getVariables().get(outVarName).getInputsForOp();
+//            if(thisOutputUsedForOps != null){
+//                Array arr = new Array(outVarName, outputFrameIter.getFrame(), outputFrameIter.getIteration(), outputFrameIter.getParentFrame());
+//                for( String opName : thisOutputUsedForOps ){
+//                    SameDiffOp o = sameDiff.getOps().get(opName);
+//                    DifferentialFunction df = o.getOp();
+//
+//                    Dep opDep;
+//                    if(df instanceof Enter) {
+//                        //Enter op: forwards input to specified frame, iteration 0
+//                        //This is a zero-copy operation
+//                        //Note that the enter value should be available for ALL iterations within the new frame - which means
+//                        // we need a frame/iter dependency, not a standard variable dependency
+//                        //Also because it's a zero-copy operation, we need a dependent alias for the input array
+//
+//                        String inName = op.getInputsToOp().get(0);
+//                        VarId inVid = lookup(inName, opInputs, allIterInputs, false);
+//                        Array inArr = new Array(inName, inVid.getFrame(), inVid.getIteration(), inVid.getParentFrame());
+//                        arrayUseTracker.addDependentAlias(inArr, arr);
+//
+//                        //Frame dependency - Enter output (and input, via alias) required for all iterations
+//                        opDep = new FrameDep(outputFrameIter.getFrame(), outputFrameIter.getParentFrame());
+//                    } else if(df instanceof Exit) {
+//                        //Exit op: forwards input to parent frame (at parent current iteration)
+//                        //This is a zero-copy operation, hence we need a dependent alias
+//                        String inName = op.getInputsToOp().get(0);
+//                        VarId inVid = lookup(inName, opInputs, allIterInputs, false);
+//                        Array inArr = new Array(inName, inVid.getFrame(), inVid.getIteration(), inVid.getParentFrame());
+//                        arrayUseTracker.addDependentAlias(inArr, arr);
+//                        //TODO not 100% sure on the opDep here...
+//                        opDep = new OpDep(opName, outputFrameIter.getFrame(), outputFrameIter.getIteration(), outputFrameIter.getParentFrame());
+//                    } else if(df instanceof NextIteration){
+//                        //NextIteration: same frame, increments iteration
+//                        //As per enter/exit etc - NextIteration is a zero-copy op, so we need a dependent alias for this
+//
+//                        String inName = op.getInputsToOp().get(0);
+//                        VarId inVid = lookup(inName, opInputs, allIterInputs, false);
+//                        Array inArr = new Array(inName, inVid.getFrame(), inVid.getIteration(), inVid.getParentFrame());
+//                        arrayUseTracker.addDependentAlias(inArr, arr);
+//
+//                        opDep = new OpDep(opName, outputFrameIter.getFrame(), outputFrameIter.getIteration(), outputFrameIter.getParentFrame());
+//                    } else {
+//                        //Normal case (standard ops) - and switch/merge cases
+//                        opDep = new OpDep(opName, outputFrameIter.getFrame(), outputFrameIter.getIteration(), outputFrameIter.getParentFrame());
+//
+//                        if( df instanceof Identity ) {
+//                            //Identity: array is passed through unchanged (same array, zero copy)
+//                            //This is fine, but we need a dependent alias. So, for arrays X -> (identity) -> Y then Y is an
+//                            // alias of X
+//
+//                            String inName = op.getInputsToOp().get(0);
+//                            Array inArr = new Array(inName, arr.getFrame(), arr.getIter(), arr.getParentFrame());
+//                            arrayUseTracker.addDependentAlias(inArr, arr);
+//                        } else if(df instanceof Switch) {
+//                            //Switch: Input to switch op is passed through unchanged (same array, zero copy) to ONE of two possible outputs
+//                            //Regardless of which branch the array is forwarded on to, the input array can be closed when the switch op is executed
+//                            //However, we need to add a dependee alias: so if we have opX -> switch -> opZ and opX has output array x
+//                            // then instead of (x -> switch) dependency, we want (x -> opZ). Otherwise, we would (incorrectly)
+//                            // say that "switch is executed, x can be deallocated"
+//
+//                            //However, we need to add a dependent alias so if opX -> switch -> opZ, and opX has output x,
+//                            // and switch has output s1 or s2, then we should mark s1 and s2 as an alias of x
+//                            //Mark it for both branches, because we don't necessarily know which will be executed at this point
+//                            Array alias0 = new Array(o.getOutputsOfOp().get(0), arr.getFrame(), arr.getIter(), arr.getParentFrame());
+//                            arrayUseTracker.addDependentAlias(arr, alias0);
+//
+//                            Array alias1 = new Array(o.getOutputsOfOp().get(1), arr.getFrame(), arr.getIter(), arr.getParentFrame());
+//                            arrayUseTracker.addDependentAlias(arr, alias1);
+//                        } else if(df instanceof Merge){
+//                            //Merge: 2 op inputs, but only one will (usually) be available when merge is executed
+//                            //The array for whichever is available is passed through unchanged (same array, zero copy)
+//                            //As per switch, we need a dependent alias - so if we have (X, Y) -> Merge -> Z
+//                            // then Z is just an alias for whichever of X or Y actually got executer
+//
+//                            //First: work out which is actually available...
+//                            String availableIn = null;
+//                            for(String s : op.getInputsToOp()){
+//                                if(constAndPhInputs != null && constAndPhInputs.contains(s)){
+//                                    availableIn = s;
+//                                    break;
+//                                }
+//                                VarId vid = lookup(s, opInputs, allIterInputs, false);
+//                                if(vid != null) {
+//                                    availableIn = s;
+//                                    break;
+//                                }
+//                            }
+//
+//                            Preconditions.checkState(availableIn != null, "Could not find any inputs for merge op %s, %s", op.getName(), outputFrameIter);
+//
+//                            Array in = new Array(availableIn, arr.getFrame(), arr.getIter(), arr.getParentFrame());
+//
+//                            Array aliasOut = new Array(o.getOutputsOfOp().get(0), arr.getFrame(), arr.getIter(), arr.getParentFrame());
+//
+////                            arrayUseTracker.addDependentAlias(in, arr);
+//                            arrayUseTracker.addDependentAlias(in, aliasOut);
+//                        }
+//                    }
+//
+//                    //Add the dependency. This means that "the specified array requires this operation to be executed, before it can be closed/deallocated"
+//                    arrayUseTracker.addDependency(arr, opDep);
+//                    log.info("Added dependency: {} -> {}", opDep, arr);
+//                }
+//            }
+//        }
+
+        //Step 2: Update dependencies - remove old dependencies (for just calculated op at given frame/iter)
+        List<String> opInVarNames = op.getInputsToOp();
+        if(opInVarNames != null){
+            //Remove any old dependencies, now that this op has been executed
+            OpDep opDep = new OpDep(op.getName(), outputFrameIter.getFrame(), outputFrameIter.getIteration(), outputFrameIter.getParentFrame());
+
+            DifferentialFunction df = op.getOp();
+            /*
+            Other control ops here:
+            - Switch: nothing special about the inputs, these can be updated as per normal ops
+            - Merge: only one was ever added as a dependency, but we can remove both (one just isn't present to be removed)
+            - Enter/Exit/NextIteration: input/output do have different frame/iter, but this should be handled by lookup method already
+             */
+
+            boolean isMerge = df instanceof Merge;
+
+
+            for( int i=0; i<opInVarNames.size(); i++ ) {
+                String n = opInVarNames.get(i);
+                VarId inVid;
+                if(constAndPhInputs != null && constAndPhInputs.contains(n)){
+                    inVid = newVarId(n, OUTER_FRAME, 0, null);
+                } else {
+                    inVid = lookup(n, opInputs, allIterInputs, !isMerge);
+                }
+
+                if(isMerge && inVid == null){
+                    //Merge op only has 1 of 2 inputs available usually
                     continue;
                 }
 
-                //Check if fully consumed (not required anywhere else)
-                boolean canDealloc = false;
-                Variable var = sameDiff.getVariables().get(inName);
-                if(var.getInputsForOp().isEmpty()){
-                    canDealloc = true;      //Should never happen
-                } else {
-                    List<String> inputForOps = var.getInputsForOp();
-                    boolean allAlreadyCalced = true;
-                    for(String opName : inputForOps){
-                        if(subgraphOps.contains(opName)){
-                            //This op IS required. We know if it has been calculated if any of the outputs of this op are present
-                            // in the set of outputs...
 
-                            //Get the VarId of the input. Note in the case of Merge(x, y) only one of x/y will be calculated
-                            SameDiffOp checkOp = sameDiff.getOps().get(opName);
+                Array arr = new Array(n, inVid.getFrame(), inVid.getIteration(), inVid.getParentFrame());
 
-                            VarId vidOfInput = lookup(inName, opInputs, allIterInputs, false);
-                            if(vidOfInput == null && !(checkOp.getOp() instanceof Merge)){
-                                //Could be nested enter case
-                                //For Enter case, this is nested enters - like constant -> enter -> enter. See getAndParameterizeOp method for this case for further details
-                                String outOfOp = sameDiff.getVariables().get(inName).getOutputOfOp();
-                                if( outOfOp != null && sameDiff.getOpById(outOfOp) instanceof Enter){
-                                    continue;
-                                }
-                                throw new IllegalStateException("Could not find VarId of input array - " + inName);
-                            }
+                //This input array no longer depends on the just executed op - it's one step closer to being able to be deallocated
+                arrayUseTracker.removeDependency(arr, opDep);
+                log.info("Removed dependency: {} -> {}", opDep, arr);
+            }
+        }
 
-                            //If currently executed op is X -> Y, and we're iterating over inputs X
-                            // then at this point opOut is Z, in X -> Z, where Z may or may not be same as Y
+        //Step 3: Check if we can deallocate any arrays
+        if(arrayUseTracker.hasZeroDependencyItem()){
+            List<Array> canDealloc = arrayUseTracker.removeAllZeroDependencyItems();
+            for(Array a : canDealloc){
 
-//                            SameDiffOp opOut = sameDiff.getOps().get(opName);
-//                            String firstOutput = opOut.getOutputsOfOp().get(0);
-                            String firstOutput = checkOp.getOutputsOfOp().get(0);
-
-
-                            if(opName.equals(op.getName())) {
-                                //Just calculated this op in this current call...
-                                allAlreadyCalced = true;
-                            } else if(checkOp.getOp() instanceof Merge){
-                                VarId vid = newVarId(firstOutput, vidOfInput.toFrameIter());
-                                allAlreadyCalced = nodeOutputs.containsKey(vid);
-                            } else if(checkOp.getOp() instanceof Switch) {
-                                //Switch has been executed if EITHER of the outputs are available. And their frame/iter is same as input
-                                VarId vid1 = newVarId(firstOutput, vidOfInput.toFrameIter());
-                                VarId vid2 = newVarId(firstOutput, vidOfInput.toFrameIter());
-
-                                allAlreadyCalced = nodeOutputs.containsKey(vid1) || nodeOutputs.containsKey(vid2);
-                            } else if(checkOp.getOp() instanceof Enter) {
-                                //Enter op: forwards input to specified execution frame
-                                /*
-                                Enter e = (Enter) checkOp.getOp();
-                                VarId vid;
-                                if(constAndPhInputs != null && constAndPhInputs.contains(inName)){
-                                    //Constant or placeholder -> always outer frame, iteration 0, no parent
-                                    vid = newVarId(firstOutput, OUTER_FRAME, 0, null);
-                                } else {
-                                    vid = newVarId(firstOutput, new FrameIter(e.getFrameName(), 0, vidOfInput.toFrameIter()));
-                                }
-                                allAlreadyCalced = nodeOutputs.containsKey(vid);
-                                System.out.println();
-                                 */
-
-                                //For enter ops: the entered value is available for multiple iterations, and is passed in as zero copy
-                                //So, we can't just close the array once the enter has triggered, unless we implemented copying
-                                // for Enter ops, because we need the array for multiple ops
-                                allAlreadyCalced = false;
-                            } else if(checkOp.getOp() instanceof Exit) {
-                                //Exit node forwards input to parent frame
-                                Exit e = (Exit) checkOp.getOp();
-                                VarId vid;
-                                if (constAndPhInputs.contains(inName)) {
-                                    //Constant or placeholder -> always outer frame, iteration 0, no parent
-                                    vid = newVarId(firstOutput, OUTER_FRAME, 0, null);
-                                } else {
-                                    FrameIter parent = outputFrameIter.getParentFrame();
-                                    vid = newVarId(firstOutput, new FrameIter(e.getFrameName(), parent.getIteration(), parent.getParentFrame()));
-                                }
-                                allAlreadyCalced = nodeOutputs.containsKey(vid);
-                            } else if(checkOp.getOp() instanceof NextIteration) {
-                                //NextIteration op: forwards its single input to the output of the current frame, but increments the iteration number
-                                FrameIter f = vidOfInput.toFrameIter().clone();
-                                f.setIteration(f.getIteration() + 1);
-                                VarId vid = newVarId(firstOutput, f);
-                                allAlreadyCalced = nodeOutputs.containsKey(vid);
-                                System.out.println(allAlreadyCalced);
-                            } else if(checkOp.getOp() instanceof Op || checkOp.getOp() instanceof DynamicCustomOp){
-                                VarId vid = newVarId(firstOutput, vidOfInput.toFrameIter());
-                                allAlreadyCalced = nodeOutputs.containsKey(vid);
-                            } else {
-                                throw new IllegalStateException("TODO: " + checkOp.getOp().getClass().getName());
-                            }
-
-
-                            /*
-                            if(opName.equals(op.getName())) {
-                                //Just calculated this op in this current call...
-                                allAlreadyCalced = true;
-                            } else if(op.getOp() instanceof Merge){
-                                VarId vid = newVarId(firstOutput, vidOfInput.toFrameIter());
-                                allAlreadyCalced = nodeOutputs.containsKey(vid);
-                            } else if(op.getOp() instanceof Switch) {
-                                //Switch has been executed if EITHER of the outputs are available. And their frame/iter is same as input
-                                VarId vid1 = newVarId(firstOutput, vidOfInput.toFrameIter());
-                                VarId vid2 = newVarId(firstOutput, vidOfInput.toFrameIter());
-
-                                allAlreadyCalced = nodeOutputs.containsKey(vid1) || nodeOutputs.containsKey(vid2);
-                            } else if(op.getOp() instanceof Enter) {
-                                //Enter op: forwards input to specified execution frame
-                                Enter e = (Enter) op.getOp();
-                                VarId vid;
-                                if(constAndPhInputs != null && constAndPhInputs.contains(inName)){
-                                    //Constant or placeholder -> always outer frame, iteration 0, no parent
-                                    vid = newVarId(firstOutput, OUTER_FRAME, 0, null);
-                                } else {
-                                    vid = newVarId(firstOutput, new FrameIter(e.getFrameName(), 0, vidOfInput.toFrameIter()));
-                                }
-                                allAlreadyCalced = nodeOutputs.containsKey(vid);
-                            } else if(op.getOp() instanceof Exit) {
-                                //Exit node forwards input to parent frame
-                                Exit e = (Exit) op.getOp();
-                                VarId vid;
-                                if (constAndPhInputs.contains(inName)) {
-                                    //Constant or placeholder -> always outer frame, iteration 0, no parent
-                                    vid = newVarId(firstOutput, OUTER_FRAME, 0, null);
-                                } else {
-                                    FrameIter parent = outputFrameIter.getParentFrame();
-                                    vid = newVarId(firstOutput, new FrameIter(e.getFrameName(), parent.getIteration(), parent.getParentFrame()));
-                                }
-                                allAlreadyCalced = nodeOutputs.containsKey(vid);
-                            } else if(op.getOp() instanceof NextIteration) {
-                                //NextIteration op: forwards its single input to the output of the current frame, but increments the iteration number
-                                FrameIter f = vidOfInput.toFrameIter().clone();
-                                f.setIteration(f.getIteration() + 1);
-                                VarId vid = newVarId(firstOutput, f);
-                                allAlreadyCalced = nodeOutputs.containsKey(vid);
-                                System.out.println(allAlreadyCalced);
-                            } else if(op.getOp() instanceof Op || op.getOp() instanceof DynamicCustomOp){
-                                VarId vid = newVarId(firstOutput, vidOfInput.toFrameIter());
-                                allAlreadyCalced = nodeOutputs.containsKey(vid);
-                            } else {
-                                throw new IllegalStateException("TODO: " + op.getOp().getClass().getName());
-                            }
-                            */
-
-                            if(!allAlreadyCalced)
-                                break;
+                //It seems like we can close "a". But we can't do that if there is an alias X == a,
+                // where X is an output we need to return to the user
+                Set<Array> s = arrayUseTracker.getDependentAliasesReverse(a);
+                if(s != null){
+                    boolean anyRequired = false;
+                    for(Array arr : s){
+                        if(allReqVariables.contains(arr.getVarName())){
+                            log.info("Not closing \"{}\" because required output \"{}\" is an alias of it", a, arr.getVarName());
+                            anyRequired = true;
+                            break;
                         }
                     }
 
-                    canDealloc = allAlreadyCalced;
-                }
-
-                if (canDealloc) {
-                    //First: Check if the array to be deallocated is the output of an enter op
-                    //Enter op arrays are available for ALL iterations, so even if it's consumed this iteration, we
-                    // should keep it for the next iteration
-                    if(var.getOutputOfOp() != null && sameDiff.getOps().get(var.getOutputOfOp()).getOp() instanceof Enter){
+                    if(anyRequired){
                         continue;
                     }
+                }
 
+                SDVariable v = sameDiff.getVariable(a.getVarName());
 
-                    INDArray arr = op.getOp().getInputArgument(i);
-
-                    /*
-                    Note regarding arrayIsConstVarOrPh(arr) call - this is to handle the case where an array is used
-                    like Constant -> enter -> enter -> switch -> X, and the output of the "switch" op isn't required
-                    anywhere else. But, for efficiency, enter/switch ops pass through the array unchanged - i.e., no
-                    copy, so the input to X is actually the exacty array for the variable "Constant" here.
-                    The "canDealloc" decision above was made locally (not knowing it's actually originally a constant),
-                     not accounting for this zero-copy behaviour
-                     */
-                    if(arr == null || arrayIsConstVarOrPh(arr))     //TODO Should null here be an error? Or OK just for merge where one may be missing?
-                        continue;
-
-
-                    log.info("Determined safe to deallocate array for: {}", inName);      //TODO FrameIter
+                if(v.getVariableType() == VariableType.ARRAY && !allReqVariables.contains(a.getVarName())){
+                    //Can't deallocate placeholders, constants, variables or arrays that the user has requested to be returned
+                    log.info("Determined safe to deallocate: {}", a);
+                    VarId vid = a.toVarId();
+                    INDArray arr = nodeOutputs.get(vid);
                     mmgr.release(arr);
-
-                    //We should also clear input references for other ops - so they aren't storing reference to closed arrays
-                    //So, if we have (x -> y) and we're closing x, we should clear x from any other ops where (x -> z)
-                    clearOpReferencesFor(inName);
-
-                    //Finally, clear the input array from nodeOutputs map (so we don't leak deallocated array reference via getArray etc)
-                    VarId vidOfInput = lookup(inName, opInputs, allIterInputs, false);
-
-                    nodeOutputs.remove(vidOfInput);
+                    clearOpReferencesFor(v.getVarName());
                 }
             }
-
         }
-
 
         return out;
     }
+
+    @Override
+    protected void recordArrayUses(List<String> forVariables, FrameIter outputFrameIter, Set<VarId> opInputs, Set<VarId> allIterInputs,
+                                   Set<String> constAndPhInputs){
+
+        for( int i=0; i<forVariables.size(); i++ ){
+            //Add the dependencies: What has to be executed, before we can close this just calculated array?
+            String outVarName = forVariables.get(i);
+            List<String> thisOutputUsedForOps = sameDiff.getVariables().get(outVarName).getInputsForOp();
+            if(thisOutputUsedForOps != null){
+                Array arr = new Array(outVarName, outputFrameIter.getFrame(), outputFrameIter.getIteration(), outputFrameIter.getParentFrame());
+                for( String opName : thisOutputUsedForOps ){
+                    SameDiffOp o = sameDiff.getOps().get(opName);
+                    DifferentialFunction df = o.getOp();
+
+                    Dep opDep;
+                    if(df instanceof Enter) {
+                        //Enter op: forwards input to specified frame, iteration 0
+                        //This is a zero-copy operation
+                        //Note that the enter value should be available for ALL iterations within the new frame - which means
+                        // we need a frame/iter dependency, not a standard variable dependency
+                        //Also because it's a zero-copy operation, we need a dependent alias for the input array
+                        String f = ((Enter) df).getFrameName();
+                        Array enterOut = new Array(o.getOutputsOfOp().get(0), f, 0, outputFrameIter);
+                        arrayUseTracker.addDependentAlias(arr, enterOut);
+
+                        //Frame dependency - Enter output (and input, via alias) required for all iterations
+                        opDep = new FrameDep(outputFrameIter.getFrame(), outputFrameIter.getParentFrame());
+                    } else if(df instanceof Exit) {
+                        //Exit op: forwards input to parent frame (at parent current iteration)
+                        //This is a zero-copy operation, hence we need a dependent alias
+                        //If arrX has just executed, and structure is arrX -> exit -> arrY, then arrY is alias for arrX
+                        String exitOutFrame = outputFrameIter.getParentFrame().getFrame();
+                        int exitOutIter = outputFrameIter.getParentFrame().getIteration();
+                        Array exitOut = new Array(o.getOutputsOfOp().get(0), exitOutFrame, exitOutIter, outputFrameIter.getParentFrame().getParentFrame());
+                        arrayUseTracker.addDependentAlias(arr, exitOut);
+                        //
+                        opDep = new OpDep(opName, outputFrameIter.getFrame(), outputFrameIter.getIteration(), outputFrameIter.getParentFrame());
+//                        opDep = new OpDep(opName, outputFrameIter.getParentFrame().getFrame(), outputFrameIter.getParentFrame().getIteration(), outputFrameIter.getParentFrame().getParentFrame());
+                    } else if(df instanceof NextIteration){
+                        //NextIteration: same frame, increments iteration
+                        //As per enter/exit etc - NextIteration is a zero-copy op, so we need a dependent alias for this
+                        //So if arrX has just executed, and structure is arrX -> nextIteration -> arrY, then arrY is an alias for arrX
+
+                        Array nextIterOut = new Array(o.getOutputsOfOp().get(0), outputFrameIter.getFrame(), outputFrameIter.getIteration()+1, outputFrameIter.getParentFrame());
+                        arrayUseTracker.addDependentAlias(arr, nextIterOut);
+
+                        opDep = new OpDep(opName, outputFrameIter.getFrame(), outputFrameIter.getIteration(), outputFrameIter.getParentFrame());
+                    } else {
+                        //Normal case (standard ops) - and switch/merge cases
+                        opDep = new OpDep(opName, outputFrameIter.getFrame(), outputFrameIter.getIteration(), outputFrameIter.getParentFrame());
+
+                        if( df instanceof Identity ) {
+                            //Identity: array is passed through unchanged (same array, zero copy)
+                            //This is fine, but we need a dependent alias. So, for arrays arrX -> (identity) -> arrY then Y is an
+                            // alias of arrX. Note that arrX is the one that has just executed
+
+                            Array identityOut = new Array(o.getOutputsOfOp().get(0), outputFrameIter.getFrame(), outputFrameIter.getIteration(), outputFrameIter.getParentFrame());
+                            arrayUseTracker.addDependentAlias(arr, identityOut);
+                        } else if(df instanceof Switch) {
+                            //Switch: Input to switch op is passed through unchanged (same array, zero copy) to ONE of two possible outputs
+                            //Regardless of which branch the array is forwarded on to, the input array can be closed when the switch op is executed
+                            //However, we need to add a dependee alias: so if we have opX -> switch -> opZ and opX has output array x
+                            // then instead of (x -> switch) dependency, we want (x -> opZ). Otherwise, we would (incorrectly)
+                            // say that "switch is executed, x can be deallocated"
+
+                            //However, we need to add a dependent alias so if opX -> switch -> opZ, and opX has output x,
+                            // and switch has output s1 or s2, then we should mark s1 and s2 as an alias of x
+                            //Mark it for both branches, because we don't necessarily know which will be executed at this point
+                            Array alias0 = new Array(o.getOutputsOfOp().get(0), arr.getFrame(), arr.getIter(), arr.getParentFrame());
+                            arrayUseTracker.addDependentAlias(arr, alias0);
+
+                            Array alias1 = new Array(o.getOutputsOfOp().get(1), arr.getFrame(), arr.getIter(), arr.getParentFrame());
+                            arrayUseTracker.addDependentAlias(arr, alias1);
+                        } else if(df instanceof Merge){
+                            //Merge: 2 op inputs, but only one will (usually) be available when merge is executed
+                            //The array for whichever is available is passed through unchanged (same array, zero copy)
+                            //As per switch, we need a dependent alias - so if we have (X, Y) -> Merge -> Z
+                            // then Z is just an alias for whichever of X or Y actually got executer
+
+                            //First: work out which is actually available...
+                            String availableIn = null;
+                            for(String s : o.getInputsToOp()){
+                                if(forVariables.contains(s)){
+                                    //Just executed op
+                                    availableIn = s;
+                                    break;
+                                }
+                                if(constAndPhInputs != null && constAndPhInputs.contains(s)){
+                                    availableIn = s;
+                                    break;
+                                }
+                                VarId vid = lookup(s, opInputs, allIterInputs, false);
+                                if(vid != null) {
+                                    availableIn = s;
+                                    break;
+                                }
+                            }
+
+                            Preconditions.checkState(availableIn != null, "Could not find any inputs for merge op %s, %s", o.getName(), outputFrameIter);
+
+                            Array in = new Array(availableIn, arr.getFrame(), arr.getIter(), arr.getParentFrame());
+
+                            Array aliasOut = new Array(o.getOutputsOfOp().get(0), arr.getFrame(), arr.getIter(), arr.getParentFrame());
+
+//                            arrayUseTracker.addDependentAlias(in, arr);
+                            arrayUseTracker.addDependentAlias(in, aliasOut);
+                        }
+                    }
+
+                    //Add the dependency. This means that "the specified array requires this operation to be executed, before it can be closed/deallocated"
+                    arrayUseTracker.addDependency(arr, opDep);
+                    log.info("Added dependency: {} -> {}", opDep, arr);
+                }
+            }
+        }
+
+    }
+
 
     public INDArray[] getOutputsHelper(DifferentialFunction op, FrameIter outputFrameIter, Set<VarId> opInputs, Set<VarId> allIterInputs,
                                        Set<String> constAndPhInputs){
@@ -402,13 +503,9 @@ public class InferenceSession extends AbstractSession<INDArray,SameDiffOp> {
             String[] argNames = i.argNames();
             Preconditions.checkState(argNames.length == 1, "Expected only 1 arg name in identity op, got %s", argNames);
             VarId vid = newVarId(argNames[0], outputFrameIter);
-            //TODO eventually we'll remove this dup for performance reasons, but no dup complicates checking for "input no longer required"
-            // because it actually *is* still
+
             INDArray orig = nodeOutputs.get(vid);
-            INDArray ret = mmgr.allocate(false, orig.dataType(), orig.shape());
-            ret.assign(orig);
-            i.setInputArgument(0, null);    //Clear reference in case it's closed later
-            return new INDArray[]{ret};
+            return new INDArray[]{orig};
         } else if(op instanceof Switch) {
             Switch s = (Switch) op;
             String[] argNames = s.argNames();       //Order: input, boolean array
@@ -1027,7 +1124,7 @@ public class InferenceSession extends AbstractSession<INDArray,SameDiffOp> {
         }
     }
 
-    protected void clearOpReferencesFor(String varName){
+    protected void clearOpReferencesFor(@NonNull String varName){
         Variable v = sameDiff.getVariables().get(varName);
 
         //Clear op outputs
@@ -1063,33 +1160,64 @@ public class InferenceSession extends AbstractSession<INDArray,SameDiffOp> {
         }
     }
 
-    /**
-     * Determine if the specified array is a constant, variable or placeholder
-     * This is used to determine when arrays are safe to deallocate
-     *
-     * See identitySetAllConstPhVar javadoc for more details
-     *
-     * @param arr Array to check
-     * @return True if constant/variable/placeholder - and false for ARRAY type SDVariable INDArray
-     */
-    protected boolean arrayIsConstVarOrPh(INDArray arr){
-        if(identitySetAllConstPhVar.isEmpty()){
-            for(SDVariable v : sameDiff.variables()){
-                if(v.getVariableType() == VariableType.PLACEHOLDER){
-                    //Only SOME placeholders may have been provided. For example, inference + no labels placeholder
-                    VarId vid = new VarId(v.getVarName(), OUTER_FRAME, 0, null);
-                    if(nodeOutputs.containsKey(vid)){
-                        identitySetAllConstPhVar.add(nodeOutputs.get(vid));
-                    }
-                } else if(v.getVariableType() != VariableType.ARRAY){
-                    //Constant or variable type
-                    identitySetAllConstPhVar.add(v.getArr());
-                }
-            }
-        }
-
-        boolean ret = identitySetAllConstPhVar.contains(arr);
-        return ret;
+    @Override
+    protected void onFrameIterTransition(String from, FrameIter parentFrom, String to, FrameIter parentTo){
+        log.info("InferenceSession2: Transition from {} (parent={}) to {} (parent={})", from, parentFrom, to, parentTo);
+        //Remove any frame dependencies...
+        //TODO
     }
 
+    @Data
+    protected static class Array {
+        private String varName;
+        private String frame;
+        private int iter;
+        private FrameIter parentFrame;
+
+        public Array(@NonNull String varName, @NonNull String frame, int iter, FrameIter parentFrame) {
+            this.varName = varName;
+            this.frame = frame;
+            this.iter = iter;
+            this.parentFrame = parentFrame;
+        }
+
+        protected VarId toVarId(){
+            return new VarId(varName, frame, iter, parentFrame);
+        }
+    }
+
+    @Data
+    protected abstract static class Dep {
+        protected String frame;
+        protected FrameIter parentFrame;
+    }
+
+    @AllArgsConstructor
+    @Data
+    @EqualsAndHashCode(callSuper = true)
+    protected static class OpDep extends Dep {
+        protected String opName;
+        protected int iter;
+
+        protected OpDep(@NonNull String opName, @NonNull String frame, int iter, FrameIter parentFrame){
+            this.opName = opName;
+            this.frame = frame;
+            this.iter = iter;
+            this.parentFrame = parentFrame;
+        }
+    }
+
+    @Data
+    @EqualsAndHashCode(callSuper = true)
+    protected static class FrameDep extends Dep {
+        public FrameDep(@NonNull String frame, FrameIter parentFrame){
+            this.frame = frame;
+            this.parentFrame = parentFrame;
+        }
+
+        @Override
+        public String toString(){
+            return "FrameDep(" + this.frame + (parentFrame == null ? "" : ",parent=" + parentFrame) + ")";
+        }
+    }
 }
